Refactor Propolis server process management into its own module
